+++ src/widget/mod.rs
@@ -32,6 +32,7 @@ pub mod roi_stats_widget;
 pub mod scalar_field_properties;
 pub mod scalar_field_view;
 pub mod scatter_mask;
+pub mod scene_position_info;
 pub mod scene_widget;
 pub mod scene_window;
 pub mod stats_widget;